requirements.txt
streamlit==1.33.0
python-dotenv==1.0.1
PyPDF2==3.0.1
langchain==0.1.16
langchain-community==0.0.32
faiss-cpu==1.8.0
huggingface-hub==0.22.2
sentence-transformers==2.7.0